requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-rest-passwordreset==1.2.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-extra-fields==3.4.0
drf-yasg==1.20.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
whitenoise==6.0.0
